requirements.odil.txt
numpy
matplotlib
odil
tensorflow-probability
tf-keras
meshio
python-box
pandas
mlflow==3.9.0
pyvista